# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AURA Dealership Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /deals:
    get:
      operationId: listDeals
      tags: [deals]
      parameters:
        - name: stage
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: List of deals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deal"
    post:
      operationId: createDeal
      tags: [deals]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
  /deals/{id}:
    get:
      operationId: getDeal
      tags: [deals]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateDeal
      tags: [deals]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /agents/runs:
    get:
      operationId: listAgentRuns
      tags: [agents]
      summary: Per-run audit records for AI agents (Admin/Leadership only)
      parameters:
        - name: agentKey
          in: query
          required: false
          schema: { type: string }
        - name: status
          in: query
          required: false
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: Recent agent runs, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentRun"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /agents/runs/{id}/review:
    post:
      operationId: reviewAgentRun
      tags: [agents]
      summary: Mark an agent run accepted or overridden (Admin/Leadership only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentRunReview"
      responses:
        "200":
          description: Updated run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRun"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /agents/{id}:
    get:
      operationId: getAgent
      tags: [agents]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateAgent
      tags: [agents]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Deal:
      type: object
      required: [id, vehicleId, stage, vehiclePrice, discount, tradeInValue, accessories, otdPrice, depositPaid, createdAt]
      properties:
        id: { type: integer }
        divisionId: { type: ["integer", "null"] }
        customerId: { type: ["integer", "null"] }
        leadId: { type: ["integer", "null"] }
        vehicleId: { type: integer }
        customerName: { type: ["string", "null"] }
        stage: { type: string, enum: [desking, negotiation, finance, committed, delivered, lost] }
        vehiclePrice: { type: number }
        discount: { type: number }
        tradeInValue: { type: number }
        accessories: { type: number }
        otdPrice: { type: number }
        monthlyPayment: { type: ["number", "null"] }
        depositPaid: { type: boolean }
        salesAdvisor: { type: ["string", "null"] }
        salesAdvisorUserId: { type: ["integer", "null"] }
        createdAt: { type: string, format: date-time }
    DealInput:
      type: object
      required: [vehicleId, vehiclePrice]
      properties:
        divisionId: { type: integer }
        customerId: { type: integer }
        leadId: { type: integer }
        vehicleId: { type: integer }
        customerName: { type: string }
        stage: { type: string, enum: [desking, negotiation, finance, committed, delivered, lost] }
        vehiclePrice: { type: number }
        discount: { type: number }
        tradeInValue: { type: number }
        accessories: { type: number }
        otdPrice: { type: number }
        monthlyPayment: { type: number }
        depositPaid: { type: boolean }
        salesAdvisor: { type: string }
        salesAdvisorUserId: { type: integer }
    DealUpdate:
      type: object
      properties:
        customerId: { type: integer }
        leadId: { type: integer }
        vehicleId: { type: integer }
        customerName: { type: string }
        stage: { type: string, enum: [desking, negotiation, finance, committed, delivered, lost] }
        vehiclePrice: { type: number }
        discount: { type: number }
        tradeInValue: { type: number }
        accessories: { type: number }
        otdPrice: { type: number }
        monthlyPayment: { type: number }
        depositPaid: { type: boolean }
        salesAdvisor: { type: string }
        salesAdvisorUserId: { type: ["integer", "null"] }
...
    Agent:
      type: object
      required: [id, key, name, domain, description, status, tasksToday, successRate]
      properties:
        id: { type: integer }
        key: { type: string }
        name: { type: string }
        domain: { type: string }
        description: { type: string }
        status: { type: string, enum: [active, idle, paused] }
        tasksToday: { type: integer }
        successRate: { type: number }
    AgentUpdate:
      type: object
      properties:
        status: { type: string, enum: [active, idle, paused] }
    AgentRun:
      type: object
      required: [id, agentKey, runType, inputSource, status, createdAt]
      properties:
        id: { type: integer }
        agentKey: { type: string }
        runType: { type: string }
        inputSource: { type: string }
        inputSummary: { type: ["string", "null"] }
        outputSummary: { type: ["string", "null"] }
        confidence: { type: ["number", "null"] }
        status:
          type: string
          enum: [completed, accepted, overridden, error, blocked]
        errorMessage: { type: ["string", "null"] }
        refType: { type: ["string", "null"] }
        refId: { type: ["integer", "null"] }
        latencyMs: { type: ["integer", "null"] }
        createdAt: { type: string, format: date-time }
    AgentRunReview:
      type: object
      required: [decision]
      properties:
        decision: { type: string, enum: [accepted, overridden] }
...