# lib/api-zod/src/generated/types/agentRun.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * AURA Dealership Operating System API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentRunStatus } from './agentRunStatus';

export interface AgentRun {
  id: number;
  agentKey: string;
  runType: string;
  inputSource: string;
  /** @nullable */
  inputSummary?: string | null;
  /** @nullable */
  outputSummary?: string | null;
  /** @nullable */
  confidence?: number | null;
  status: AgentRunStatus;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  refType?: string | null;
  /** @nullable */
  refId?: number | null;
  /** @nullable */
  latencyMs?: number | null;
  createdAt: Date;
}