# lib/api-zod/src/generated/types/agentRunReview.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * AURA Dealership Operating System API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentRunReviewDecision } from './agentRunReviewDecision';

export interface AgentRunReview {
  decision: AgentRunReviewDecision;
}